:hammer: Refactor theme implementation
