+++ composeApp/src/commonMain/kotlin/com/crosspaste/ui/ThemeBackground.kt
@@ -8,6 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
+import com.crosspaste.ui.CrossPasteTheme.isLight
 
 @Composable
 fun ThemeBackground() {
