+++ composeApp/src/commonMain/kotlin/com/crosspaste/ui/ThemeDetector.kt
@@ -1,0 +1,18 @@
+package com.crosspaste.ui
+
+interface ThemeDetector {
+    fun isSystemInDark(): Boolean
+
+    fun isFollowSystem(): Boolean
+
+    fun isUserInDark(): Boolean
+
+    fun isCurrentThemeDark(): Boolean {
+        return isFollowSystem() && isSystemInDark() || !isFollowSystem() && isUserInDark()
+    }
+
+    fun setThemeConfig(
+        isFollowSystem: Boolean,
+        isUserInDark: Boolean = false,
+    )
+}
