+++ composeApp/src/commonMain/kotlin/com/crosspaste/ui/base/MenuView.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.crosspaste.ui.selectColor
+import com.crosspaste.ui.CrossPasteTheme.selectColor
 
 val menuItemTextStyle =
     TextStyle(
