+++ composeApp/src/desktopMain/kotlin/com/crosspaste/ui/WindowsTrayView.kt
@@ -34,6 +34,7 @@ import com.crosspaste.app.DesktopAppWindowManager
 import com.crosspaste.app.WinAppWindowManager
 import com.crosspaste.composeapp.generated.resources.Res
 import com.crosspaste.composeapp.generated.resources.crosspaste
+import com.crosspaste.ui.CrossPasteTheme.Theme
 import com.crosspaste.ui.base.DesktopNotificationManager
 import com.crosspaste.ui.base.NotificationManager
 import com.crosspaste.utils.GlobalCoroutineScopeImpl.mainCoroutineDispatcher
@@ -149,7 +150,7 @@ object WindowsTrayView {
         val appSize = koinInject<AppSize>() as DesktopAppSize
         val appWindowManager = koinInject<DesktopAppWindowManager>()
 
-        CrossPasteTheme {
+        Theme {
             Box(
                 modifier =
                     Modifier
