+++ composeApp/src/desktopMain/kotlin/com/crosspaste/ui/devices/DeviceConnectView.desktop.kt
@@ -54,6 +54,7 @@ import com.crosspaste.i18n.GlobalCopywriter
 import com.crosspaste.realm.sync.SyncRuntimeInfo
 import com.crosspaste.realm.sync.SyncState
 import com.crosspaste.sync.SyncManager
+import com.crosspaste.ui.CrossPasteTheme.selectColor
 import com.crosspaste.ui.ScreenType
 import com.crosspaste.ui.base.MenuItem
 import com.crosspaste.ui.base.MessageType
@@ -61,7 +62,6 @@ import com.crosspaste.ui.base.NotificationManager
 import com.crosspaste.ui.base.PasteIconButton
 import com.crosspaste.ui.base.getMenWidth
 import com.crosspaste.ui.base.moreVertical
-import com.crosspaste.ui.selectColor
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import org.koin.compose.koinInject
