+++ composeApp/src/desktopMain/kotlin/com/crosspaste/ui/devices/SyncDeviceView.desktop.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.input.pointer.onPointerEvent
 import androidx.compose.ui.unit.dp
 import com.crosspaste.dto.sync.SyncInfo
 import com.crosspaste.realm.sync.createSyncRuntimeInfo
-import com.crosspaste.ui.selectColor
+import com.crosspaste.ui.CrossPasteTheme.selectColor
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
